Clarify what/why/how
Update README to clarify what TagBot does, why its neat, and what users need to do after adding the `TagBot.yml` file. What I wrote is based on my understanding of TagBot - I never used the prior version of it and I thought that the README was fuzzy on the parts that I tried to clarify. It may be that I'm still not clear on these items but would love to modify my pull request if that's the case.