Clone private registries over SSH
The end goal here is to be able to clone a private registry with a
read-only deploy key for that registry repo.

So this actually won't work (currently) as the SSH key it uses is
almost always a write-enabled deploy key for the package repo, which
itself won't have access to the registry.

This is going to take on a lot of technical debt... I really did not
design with this in mind. This, and especially subdir support,
probably warrants a substantial redesign.